conf/config.jsファイルを作成し、db設定内容を分離
