+++ routes/login/login.js
@@ -10,22 +10,7 @@ let pg = require('pg');
 let Pool = require('pg-pool');
 let dotenv = require('dotenv').config();
 
-const url = require('url');
-
-const params = url.parse(process.env.DATABASE_URL);
-const auth = params.auth.split(':');
-
-const db_config = {
-      host:  params.hostname,
-      database: params.pathname.split('/')[1],
-      user: auth[0],
-      password: auth[1],
-      port: params.port,
-      ssl: true,
-      max: 10, //set pool max size to 20
-      min: 4, //set min pool size to 4
-      idleTimeoutMillis: 1000 //close idle clients after 1 second
-};
+const config = require('../../conf/config');
 
 /* GET login listing. */
 router.get('/', function(req, res) {
@@ -56,7 +41,7 @@ router.post('/', function(req, res, next) {
 
     }else{
 
-        let pool = new Pool(db_config);
+        let pool = new Pool(config.db_config);
 
         pool.connect()
